refactor: move all validation in POST /work_experiences and /interview_experiences to graphql layer
