+++ schema.graphql
@@ -19,7 +19,7 @@ type Company {
 
 input CompanyInput {
   id: String
-  name: String
+  query: String!
 }
 
 input CreateInterviewExperienceInput {
@@ -31,14 +31,14 @@ input CreateInterviewExperienceInput {
   sections: [SectionInput!]!
   experience_in_year: Int
   education: String
-  status: String
+  status: String = "published"
   email: String
 
   """interview part"""
   interview_time: InterviewTimeInput!
   interview_result: String!
-  interview_qas: [InterviewQuestionInput!]
-  interview_sensitive_questions: [String!]
+  interview_qas: [InterviewQuestionInput!] = []
+  interview_sensitive_questions: [String!] = []
   salary: SalaryInput
   overall_rating: Int!
 }
@@ -57,17 +57,17 @@ input CreateWorkExperienceInput {
   sections: [SectionInput!]!
   experience_in_year: Int
   education: String
-  status: String
+  status: String = "published"
   email: String
 
   """work part"""
   salary: SalaryInput
   week_work_time: Int
-  recommend_to_others: recommendToOthersType
-  is_currently_employed: isCurrentEmployedType!
+  recommend_to_others: RecommendToOthersType
+  is_currently_employed: IsCurrentEmployedType!
 
   """will have this column if 'is_currently_employed' === no"""
-  job_ending_time: jobEndingTimeInput
+  job_ending_time: JobEndingTimeInput
 }
 
 type CreateWorkExperiencePayload {
@@ -196,12 +196,12 @@ type InterviewExperienceStatistics {
 }
 
 type InterviewQuestion {
-  question: String
+  question: String!
   answer: String
 }
 
 input InterviewQuestionInput {
-  question: String
+  question: String!
   answer: String
 }
 
@@ -210,7 +210,7 @@ input InterviewTimeInput {
   month: Int!
 }
 
-enum isCurrentEmployedType {
+enum IsCurrentEmployedType {
   yes
   no
 }
@@ -221,7 +221,7 @@ type JobAverageSalary {
   data_count: Int!
 }
 
-input jobEndingTimeInput {
+input JobEndingTimeInput {
   year: Int!
   month: Int!
 }
@@ -337,7 +337,7 @@ type Query {
   salary_work_time_count: Int!
 }
 
-enum recommendToOthersType {
+enum RecommendToOthersType {
   yes
   no
 }
@@ -380,7 +380,7 @@ type SalaryDistributionBin {
 }
 
 input SalaryInput {
-  type: String!
+  type: SalaryType!
   amount: Float!
 }
 
@@ -441,7 +441,7 @@ enum SearchBy {
 
 type Section {
   subtitle: String
-  content: String
+  content: String!
 }
 
 input SectionInput {
