+++ src/schema/experience.js
@@ -1,17 +1,190 @@
 const { gql, UserInputError } = require("apollo-server-express");
 const { ObjectId } = require("mongodb");
 const R = require("ramda");
-const winston = require("winston");
 
-const { requiredNumberInRange } = require("../libs/validation");
+const { HttpError } = require("../libs/errors");
+const {
+    shouldIn,
+    stringRequireLength,
+    validateEmail,
+    requiredNumberInRange,
+} = require("../libs/validation");
 const ExperienceModel = require("../models/experience_model");
 const UserModel = require("../models/user_model");
+const helper = require("../routes/company_helper");
 
 const WorkExperienceType = "work";
 const InterviewExperienceType = "interview";
 const InternExperienceType = "intern";
 
 const MAX_PREVIEW_SIZE = 160;
+
+function validateCommonInputFields(data) {
+    if (
+        !shouldIn(data.region, [
+            "彰化縣",
+            "嘉義市",
+            "嘉義縣",
+            "新竹市",
+            "新竹縣",
+            "花蓮縣",
+            "高雄市",
+            "基隆市",
+            "金門縣",
+            "連江縣",
+            "苗栗縣",
+            "南投縣",
+            "新北市",
+            "澎湖縣",
+            "屏東縣",
+            "臺中市",
+            "臺南市",
+            "臺北市",
+            "臺東縣",
+            "桃園市",
+            "宜蘭縣",
+            "雲林縣",
+        ])
+    ) {
+        throw new HttpError(`地區不允許 ${data.region}！`, 422);
+    }
+
+    if (!stringRequireLength(data.title, 1, 50)) {
+        throw new HttpError("標題僅限 1~50 字！", 422);
+    }
+
+    if (!data.sections || !(data.sections instanceof Array)) {
+        throw new HttpError("內容要寫喔！", 422);
+    }
+    data.sections.forEach(section => {
+        if (
+            section.subtitle !== null &&
+            !stringRequireLength(section.subtitle, 1, 25)
+        ) {
+            throw new HttpError("內容標題僅限 1~25 字！", 422);
+        }
+        if (!stringRequireLength(section.content, 1, 5000)) {
+            throw new HttpError("內容標題僅限 1~5000 字！", 422);
+        }
+    });
+
+    if (data.experience_in_year) {
+        if (data.experience_in_year < 0 || data.experience_in_year > 50) {
+            throw new HttpError("相關職務工作經驗需大於等於0，小於等於50", 422);
+        }
+    }
+
+    if (data.education) {
+        if (
+            !shouldIn(data.education, [
+                "大學",
+                "碩士",
+                "博士",
+                "高職",
+                "五專",
+                "二專",
+                "二技",
+                "高中",
+                "國中",
+                "國小",
+            ])
+        ) {
+            throw new HttpError("最高學歷範圍錯誤", 422);
+        }
+    }
+
+    if (data.email && !validateEmail(data.email)) {
+        throw new HttpError("E-mail 格式錯誤", 422);
+    }
+
+    if (data.salary && data.salary.amount < 0) {
+        throw new HttpError("薪資不小於0", 422);
+    }
+}
+
+function validateWorkInputFields(data) {
+    if (data.is_currently_employed === "no") {
+        if (!data.job_ending_time) {
+            throw new HttpError("離職年、月份要填喔！", 422);
+        }
+        const now = new Date();
+        if (data.job_ending_time.year <= now.getFullYear() - 10) {
+            throw new HttpError("離職年份需在10年內", 422);
+        }
+        if (data.job_ending_time.month < 1 || data.job_ending_time.month > 12) {
+            throw new HttpError("離職月份需在1~12月", 422);
+        }
+        if (
+            (data.job_ending_time.year === now.getFullYear() &&
+                data.job_ending_time.month > now.getMonth() + 1) ||
+            data.job_ending_time.year > now.getFullYear()
+        ) {
+            throw new HttpError("離職月份不可能比現在時間晚", 422);
+        }
+    }
+
+    if (data.week_work_time) {
+        if (data.week_work_time < 0 || data.week_work_time > 168) {
+            throw new HttpError("工時需介於 0~168 之間", 422);
+        }
+    }
+}
+
+function validationWorkInputFields(data) {
+    validateCommonInputFields(data);
+    validateWorkInputFields(data);
+}
+
+function validateInterviewInputFields(data) {
+    const now = new Date();
+    if (data.interview_time.year <= now.getFullYear() - 10) {
+        throw new HttpError("面試年份需在10年內", 422);
+    }
+    if (data.interview_time.month < 1 || data.interview_time.month > 12) {
+        throw new HttpError("面試月份需在1~12月", 422);
+    }
+    if (
+        (data.interview_time.year === now.getFullYear() &&
+            data.interview_time.month > now.getMonth() + 1) ||
+        data.interview_time.year > now.getFullYear()
+    ) {
+        throw new HttpError("面試月份不可能比現在時間晚", 422);
+    }
+
+    data.interview_qas.forEach(qa => {
+        if (!stringRequireLength(qa.question, 1, 250)) {
+            throw new HttpError("面試題目標題僅限 1~250 字！", 422);
+        }
+
+        if (qa.answer) {
+            if (!stringRequireLength(qa.answer, 1, 5000)) {
+                throw new HttpError("面試題目標題僅限 1~5000 字！", 422);
+            }
+        }
+    });
+    if (data.interview_qas.length > 30) {
+        throw new HttpError("面試題目列表超過 30 題！", 422);
+    }
+
+    if (!stringRequireLength(data.interview_result, 1, 100)) {
+        throw new HttpError("面試結果僅限 1~100 字！", 422);
+    }
+
+    data.interview_sensitive_questions.forEach(question => {
+        if (!stringRequireLength(question, 1, 20)) {
+            throw new HttpError("面試中提及的特別問題僅限 1~20 字！", 422);
+        }
+    });
+
+    if (!shouldIn(data.overall_rating, [1, 2, 3, 4, 5])) {
+        throw new HttpError("面試分數有誤", 422);
+    }
+}
+
+function validationInterviewInputFields(data) {
+    validateCommonInputFields(data);
+    validateInterviewInputFields(data);
+}
 
 const Type = gql`
     interface Experience {
@@ -150,11 +323,11 @@ const Type = gql`
 
     type Section {
         subtitle: String
-        content: String
+        content: String!
     }
 
     type InterviewQuestion {
-        question: String
+        question: String!
         answer: String
     }
 `;
@@ -171,7 +344,7 @@ const Query = gql`
     }
 `;
 
-const Mutation = `
+const Mutation = gql`
     input CreateInterviewExperienceInput {
         "Common"
         company: CompanyInput!
@@ -181,20 +354,20 @@ const Mutation = `
         sections: [SectionInput!]!
         experience_in_year: Int
         education: String
-        status: String = published
+        status: String = "published"
         email: String
         "interview part"
         interview_time: InterviewTimeInput!
         interview_result: String!
-        interview_qas: [InterviewQuestionInput!]
-        interview_sensitive_questions: [String!]
+        interview_qas: [InterviewQuestionInput!] = []
+        interview_sensitive_questions: [String!] = []
         salary: SalaryInput
         overall_rating: Int!
     }
 
     input CompanyInput {
         id: String
-        name: String
+        query: String!
     }
 
     input InterviewTimeInput {
@@ -203,7 +376,7 @@ const Mutation = `
     }
 
     input SalaryInput {
-        type: String!
+        type: SalaryType!
         amount: Float!
     }
 
@@ -213,7 +386,7 @@ const Mutation = `
     }
 
     input InterviewQuestionInput {
-        question: String
+        question: String!
         answer: String
     }
 
@@ -231,28 +404,28 @@ const Mutation = `
         sections: [SectionInput!]!
         experience_in_year: Int
         education: String
-        status: String = published
+        status: String = "published"
         email: String
         "work part"
         salary: SalaryInput
         week_work_time: Int
-        recommend_to_others: recommendToOthersType
-        is_currently_employed: isCurrentEmployedType!
+        recommend_to_others: RecommendToOthersType
+        is_currently_employed: IsCurrentEmployedType!
         "will have this column if 'is_currently_employed' === no"
-        job_ending_time: jobEndingTimeInput
+        job_ending_time: JobEndingTimeInput
     }
 
-    enum isCurrentEmployedType {
+    enum IsCurrentEmployedType {
         yes
         no
     }
 
-    enum recommendToOthersType {
+    enum RecommendToOthersType {
         yes
         no
     }
 
-    input jobEndingTimeInput {
+    input JobEndingTimeInput {
         year: Int!
         month: Int!
     }
@@ -419,9 +592,31 @@ const resolvers = {
         async createInterviewExperience(
             root,
             { input },
-            { db, user, manager, ip, ips }
+            { db, user, manager }
         ) {
             const experience = input;
+
+            validationInterviewInputFields(experience);
+
+            const company_model = manager.CompanyModel;
+
+            const company = await helper.getCompanyByIdOrQuery(
+                company_model,
+                input.company.id,
+                input.company.query
+            );
+            experience.company = company;
+            experience.job_title = experience.job_title.toUpperCase();
+            experience.interview_qas = experience.interview_qas.map(qas => {
+                const result = {
+                    question: qas.question,
+                };
+                if (typeof qas.answer === "undefined" || qas.answer == null) {
+                    return result;
+                }
+                result.answer = qas.answer;
+                return result;
+            });
 
             Object.assign(experience, {
                 type: "interview",
@@ -452,23 +647,25 @@ const resolvers = {
                 );
             }
 
-            winston.info("interview experiences insert data success", {
-                id: experience._id,
-                ip,
-                ips,
-            });
-
             return {
                 success: true,
                 experience,
             };
         },
-        async createWorkExperience(
-            root,
-            { input },
-            { db, user, manager, ip, ips }
-        ) {
+        async createWorkExperience(root, { input }, { db, user, manager }) {
             const experience = input;
+
+            validationWorkInputFields(experience);
+
+            const company_model = manager.CompanyModel;
+
+            const company = await helper.getCompanyByIdOrQuery(
+                company_model,
+                input.company.id,
+                input.company.query
+            );
+            experience.company = company;
+            experience.job_title = experience.job_title.toUpperCase();
 
             Object.assign(experience, {
                 type: "work",
@@ -507,12 +704,6 @@ const resolvers = {
                     experience.email
                 );
             }
-
-            winston.info("work experiences insert data success", {
-                id: experience._id,
-                ip,
-                ips,
-            });
 
             return {
                 success: true,
